+++ app/shared/src/commonMain/kotlin/session/SessionEvent.kt
@@ -1,0 +1,12 @@
+package me.him188.ani.app.session
+
+sealed interface SessionEvent {
+    /**
+     * token 有变更
+     */
+    sealed interface UserActionEvent : SessionEvent
+    data object Login : UserActionEvent
+    data object Logout : UserActionEvent
+
+    data object TokenRefreshed : SessionEvent
+}
